Refactored throughputwrapper to be component agnostic
