+++ internal/throughputwrapper/metric_consumer_test.go
