+++ internal/throughputwrapper/metrics.go
@@ -20,38 +20,38 @@ import (
 	"go.opencensus.io/tag"
 )
 
-const tagReceiverKey = "receiver"
+const tagComponentKey = "component"
 
 var (
-	receiverTagKey       = tag.MustNewKey(tagReceiverKey)
-	logThroughputSize    = stats.Int64("log_throughput_size", "Size of the log package emitted from the receiver", stats.UnitBytes)
-	metricThroughputSize = stats.Int64("metric_throughput_size", "Size of the metric package emitted from the receiver", stats.UnitBytes)
-	traceThroughputSize  = stats.Int64("trace_throughput_size", "Size of the trace package emitted from the receiver", stats.UnitBytes)
+	componentTagKey      = tag.MustNewKey(tagComponentKey)
+	logThroughputSize    = stats.Int64("log_throughput_size", "Size of the log package emitted from the component", stats.UnitBytes)
+	metricThroughputSize = stats.Int64("metric_throughput_size", "Size of the metric package emitted from the component", stats.UnitBytes)
+	traceThroughputSize  = stats.Int64("trace_throughput_size", "Size of the trace package emitted from the component", stats.UnitBytes)
 )
 
 func metricViews() []*view.View {
-	receiverTagKeys := []tag.Key{receiverTagKey}
+	componentTagKeys := []tag.Key{componentTagKey}
 
 	return []*view.View{
 		{
-			Name:        tagReceiverKey + "/" + logThroughputSize.Name(),
+			Name:        tagComponentKey + "/" + logThroughputSize.Name(),
 			Description: logThroughputSize.Description(),
 			Measure:     logThroughputSize,
-			TagKeys:     receiverTagKeys,
+			TagKeys:     componentTagKeys,
 			Aggregation: view.Sum(),
 		},
 		{
-			Name:        tagReceiverKey + "/" + metricThroughputSize.Name(),
+			Name:        tagComponentKey + "/" + metricThroughputSize.Name(),
 			Description: metricThroughputSize.Description(),
 			Measure:     metricThroughputSize,
-			TagKeys:     receiverTagKeys,
+			TagKeys:     componentTagKeys,
 			Aggregation: view.Sum(),
 		},
 		{
-			Name:        tagReceiverKey + "/" + traceThroughputSize.Name(),
+			Name:        tagComponentKey + "/" + traceThroughputSize.Name(),
 			Description: traceThroughputSize.Description(),
 			Measure:     traceThroughputSize,
-			TagKeys:     receiverTagKeys,
+			TagKeys:     componentTagKeys,
 			Aggregation: view.Sum(),
 		},
 	}
