+++ internal/throughputwrapper/trace_consumer_test.go
